refactor: Constants to enums, del Category enum
Use enums instead of constants and remove
Category enum.

Based on the new enums it's possible to
determine the categories.

Rosh Chodesh is a redundant category and
is now under Chag.

The constant dicts for Israel and Diaspora
have been replaced by dynamically building
the months. This improves code readability
and even speeds up the code.
Ran 46 tests in 0.257s
Ran 49 tests in 0.217s